+++ src/app/frontend/common/components/resourcecard/logsbutton_component.js
